✨ introduce specialists (#336)
This introduces the notion of some users being specialists in some subjects (aka tag labels). Specialists will have more permissions on questions and answers which are tagged with their specialties, though no permissions are introduced here. The specialties of a user are displayed on their profile. This does not introduce means to define specialties through the UI, this will come later.

See also #334.