:bug: change static to assets in staticPaths (#351)
Vite changed the folder of built assets (CSS and JS files) from `static` to `assets` (see #346). However, the production server treated `static` specially as a folder where it should not serve `index.html`. As a result, the server served `index.html` for all files located in `assets`, and assets could not be served in production. This fix adds `assets` to the list of files and folders that are treated specially by the server (in place of `static`).